% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IS.R
\name{IS}
\alias{IS}
\title{Interdaily Statbility}
\usage{
IS(x)
}
\arguments{
\item{x}{\code{data.frame} of dimension ndays by p, where p is the dimension of the data.}
}
\description{
This function calcualte interdaily stability, a nonparametric metric
of circadian rhtymicity
}
\examples{
data(example_activity_data)
count1 = example_activity_data$count[c(1,2,3),-c(1,2)]
is = IS(x = count1)
}
\references{
Junrui Di et al. Joint and individual representation of domains of physical activity, sleep, and circadian rhythmicity. Statistics in Biosciences.
}
