% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IS_long.R
\name{IS_long}
\alias{IS_long}
\title{Interdaily Statbility for the Whole Dataset}
\usage{
IS_long(count.data, window = 1)
}
\arguments{
\item{count.data}{\code{data.frame} of dimension n * (p+2) containing the
p dimensional activity data for all n subject days.
The first two columns have to be ID and Day. ID can be
either \code{character} or \code{numeric}. Day has to be \code{numeric} indicating
the sequency of days within each subject.}

\item{window}{an \code{integer} indicating what is the window to bin the data before
the function can be apply to the dataset. For details, see \code{bin_data}.}
}
\value{
IV
}
\description{
This function calcualte interdaily stability, a nonparametric metric
of circadian rhtymicity. This function is a whole dataset
wrapper for \code{IS}
}
\examples{
data(example_activity_data)
count1 = example_activity_data$count
is_subj = IS_long(count.data = count1, window = 10)
}
\references{
Junrui Di et al. Joint and individual representation of domains of physical activity, sleep, and circadian rhythmicity. Statistics in Biosciences.
}
