% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IV.R
\name{IV}
\alias{IV}
\title{Intradaily Variability}
\usage{
IV(x)
}
\arguments{
\item{x}{\code{vector} of activity data}
}
\value{
IV
}
\description{
This function calcualte intradaily variability, a nonparametric metric
reprsenting fragmentation of circadian rhtymicity
}
\examples{
data(example_activity_data)
count1 = c(t(example_activity_data$count[1,-c(1,2)]))
iv = IV(x = count1)


}
\references{
Junrui Di et al. Joint and individual representation of domains of physical activity, sleep, and circadian rhythmicity. Statistics in Biosciences.
}
