% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IV_long.R
\name{IV_long}
\alias{IV_long}
\title{Intradaily Variability for the Whole Dataset}
\usage{
IV_long(count.data, window = 1, method = c("average", "sum"))
}
\arguments{
\item{count.data}{\code{data.frame} of dimension n * (p+2) containing the
p dimensional activity data for all n subject days.
The first two columns have to be ID and Day. ID can be
either \code{character} or \code{numeric}. Day has to be \code{numeric} indicating
the sequency of days within each subject.}

\item{window}{an \code{integer} indicating what is the window to bin the data before
the function can be apply to the dataset. For details, see \code{bin_data}.}

\item{method}{\code{character} of "sum" or "average", function used to bin the data}
}
\value{
IV
}
\description{
This function calcualte intradaily variability, a nonparametric metric
reprsenting fragmentation of circadian rhtymicity. This function is a whole dataset
wrapper for \code{IV}.
}
\examples{
data(example_activity_data)
count1 = example_activity_data$count
iv_subj = IV_long(count.data = count1, window = 10, method = "average")

}
\references{
Junrui Di et al. Joint and individual representation of domains of physical activity, sleep, and circadian rhythmicity. Statistics in Biosciences.
}
