% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RA.R
\name{RA}
\alias{RA}
\title{Relative Amplitude}
\usage{
RA(x, window = 1, method = c("average", "sum"))
}
\arguments{
\item{x}{\code{vector} vector of activity data}

\item{window}{since the caculation of M10 and L5 depends on the dimension of data, we need to include
window size as an argument.}

\item{method}{\code{character} of "sum" or "average", function used to bin the data}
}
\value{
RA
}
\description{
This function calcualte relative amplitude, a nonparametric metric
reprsenting fragmentation of circadian rhtymicity
}
\examples{
data(example_activity_data)
count1 = c(t(example_activity_data$count[1,-c(1,2)]))
ra = RA(x = count1, window = 10, method = "average")


}
\references{
Junrui Di et al. Joint and individual representation of domains of physical activity, sleep, and circadian rhythmicity. Statistics in Biosciences.
}
